[core]: fix namelen size in case it is negative

+++ srtcore/api.cpp
@@ -1179,7 +1179,7 @@ SRTSOCKET srt::CUDTUnited::accept(const SRTSOCKET listen, sockaddr* pw_addr, int
 int srt::CUDTUnited::connect(SRTSOCKET u, const sockaddr* srcname, const sockaddr* tarname, int namelen)
 {
     // Here both srcname and tarname must be specified
-    if (!srcname || !tarname || size_t(namelen) < sizeof(sockaddr_in))
+    if (!srcname || !tarname || namelen < int(sizeof(sockaddr_in)))
     {
         LOGC(aclog.Error,
              log << "connect(with source): invalid call: srcname=" << srcname << " tarname=" << tarname
@@ -1224,7 +1224,7 @@ int srt::CUDTUnited::connect(SRTSOCKET u, const sockaddr* srcname, const sockadd
 
 int srt::CUDTUnited::connect(const SRTSOCKET u, const sockaddr* name, int namelen, int32_t forced_isn)
 {
-    if (!name || size_t(namelen) < sizeof(sockaddr_in))
+    if (!name || namelen < int(sizeof(sockaddr_in)))
     {
         LOGC(aclog.Error, log << "connect(): invalid call: name=" << name << " namelen=" << namelen);
         throw CUDTException(MJ_NOTSUP, MN_INVAL);
